requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
google-genai>=1.0.0
chromadb>=0.6.0
python-dotenv>=1.0.0
